extract jit kernels
